Fix MTR for builds without performance schema
Change `--disable-performance-schema` to `--loose-disable-performance-schema`.

Sergei Golubchik broke this in 2bc940f7c940. See
https://github.com/MariaDB/server/commit/2bc940f7c940#r140575671.

It was subsequently fixed in
https://github.com/MariaDB/server/commit/354e97cd722c6b8721f05e75084a6843e32b6b74,
but that was never merged to 11.3 / 11.4 / 11.5.

All new code of the whole pull request, including one or several files
that are either new files or modified ones, are contributed under the
BSD-new license. I am contributing on behalf of my employer Amazon Web
Services, Inc.