MDEV-32020: Replicate XA PREPARE/COMMIT/ROLLBACK on the slave
Implement that slave preserves binlogs of pending prepared XA transactions.
Implement that slave records, in a crash-safe way, those prepared trx in the
binlog that are not yet applied in engine, in preparation for implementing
XA RECOVER and XA PREPARE/COMMIT/ROLLBACK of pending XA prepared trx when
promoting a slave to a master.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>