Add option to suppress repeated error log messages
This change adds a new configuration option
`log_suppress_repeated_errors` that allows users to control the behavior
of error logging. When this option is set to a non-zero value, the
server will suppress repeated identical error messages after the
specified number of occurrences.

This is implemented by tracking the last logged error message and its
count. If a new message matches the previous one, and the count has not
reached the suppression limit, the message is not logged again. Once a
new unique message is encountered, the previous message's count is
logged, and the new message is printed.

The suppression can be disabled by setting the option to 0.

```
2024-04-25 21:38:03 0 [Note] /usr/local/mysql/bin/mariadbd: ready for connections.
Version: '11.4.2-MariaDB'  socket: '/tmp/mysql.sock'  port: 3306  Source distribution
2024-04-25 21:38:11 3 [Warning] Access denied for user 'foo'@'localhost' (using password: YES)
2024-04-25 21:38:28 8 [Warning] Previous error was emitted 5 times
2024-04-25 21:38:28 8 [Warning] Access denied for user 'bar'@'localhost' (using password: YES)
2024-04-25 21:38:36 13 [Warning] Previous error was emitted 5 times
2024-04-25 21:38:36 13 [Warning] Access denied for user 'foo'@'localhost' (using password: YES)
2024-04-25 21:38:41 14 [Warning] Access denied for user 'bar'@'localhost' (using password: YES)
```