Use clang for GitLab sanitizers and allow MTR overrides
Many sanitizer violations that are outputted are compiler
specific. Running the MTR with sanitizers enabled when compiling
with clang:latest will still output violations that are not
experienced when building with gcc. This causes an ever-failing
build which makes it harder to detect new genuine sanitizer
related regressions. Refactor GitLab sanitizer jobs to use gcc
instead.

Additionally, the GitLab MTR job is subject to many MTR default
parameters. One of them being the max test failure amount of 10.
Refactor the MTR definition to allow for job specific overrides.
In doing the above we allow for the sanitizer jobs to run with
failures uncapped to output all violations for the entire MTR
main suite as jobs like TSAN currently do not run the
MTR without violations and are allowed to fail anyways.

All new code of the whole pull request, including one or several
files that are either new files or modified ones, are contributed
under the BSD-new license. I am contributing on behalf of my
employer Amazon Web Services, Inc.